Fixed a bug in the pruning algorithmus which caused some Policies to not get pruned correctly or only relations to get removed

+++ TD-Game/src/nl/uva/td/ai/mcts/SearchTree.java
@@ -15,7 +15,7 @@ public class SearchTree {
     /** The total amount of nodes in this search tree **/
     private long mNodeCounter = 0;
 
-    private long mCurrentDeepestDepth = 0;
+    private int mCurrentDeepestDepth = 0;
 
     /**
      * Initialise the search tree with the root node
